Patch some bugs and new feature
There's a ini settings that is on false actually, if you put it on true, it will ask you which serie is the right. On false it will determine by himself the best one, it can cause some problem cause of the levenshtein algorithm. I patched a bug where the language setting doesn't do anything, it's resolved.